requirement.txt
asgiref==3.5.1
Babel==2.10.3
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
numpy==1.22.3
opencv-python==4.6.0.66
pandas==1.4.2
pika==1.2.1
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2022.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
torch==1.12.1
torchvision==0.13.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.12
